backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-genai>=0.3.0
python-multipart>=0.0.9
networkx>=3.2.1
chromadb>=0.4.22
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
requests>=2.31.0
